+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 * Fixed an exception when adding custom fields.
+* The format of outgoing timestamp fields can now be customised in the session properties dialog.